Add form input capability
Now the user can enter a book in the form that's going to be
displayed in the table. The process of doing so has also been simplified,
by using a myLibrary array containing book objects, every time the user
enters a book, it is added to that array, and everytime the submit
button is pressed, the HTML table is updated with the books
in myLibrary. I also started working on the close capability,
which will allow the user to remove a book from the table.