Improve remove book functionality
Now for the first time, books can be added and removed in any combination
without bugs and the html table stays synced with 'myLibrary'.
This is thanks to an idea that was important: Every book in the html
table has a unique id (in the <td> that has the remove button), that
matches the id of the book object. When the remove button is clicked,
the program looks for the id of the <td> and, while removing
the parent <tr> from the table, it also removes the book with the matching
id in 'myLibrary', thanks to the findIndex() function.