Add thead on/off visibility functionality
Now the thead appears on screen only when there are books in the
table. This is thanks to toggling the display on and off with JS.
An important addition to this feature was to divide the thead and
the tbody on the HTML, and the function addBookToTable() to inject
the HTML code on the tbody instead of the whole table, which caused
problems to the event handlers needed to achieve this feature.