Add book constructor and toggle read function
The books are now created with the 'Book' constructor function.
This is also useful because it allows the 'toggleRead' function to
be added to the prototype of the constructor. Some other parts of the
code have also been optimized such as if statements written as ternary
operator and improved comments