Add JS form form validation
If any of the input values are missing, the 'show' class will be toggled,
showing the error messages in their corresponding inputs. Only when all
the inputs of the form are valid (with .checkValidity()), the
library object will be updated with the new data and renderLibrary()
will be called. The pages input has also been updated to 'number' type,
effectively preventing the user to type non-numbers